% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_renter_share.R
\name{build_coc_evictions}
\alias{build_coc_evictions}
\title{Creates eviction data}
\usage{
build_coc_evictions(evictions, county_crosswalk)
}
\arguments{
\item{evictions}{A data frame with county-level eviction data.}

\item{county_crosswalk}{A data frame with the county to CoC crosswalk.}
}
\value{
A data frame:
\itemize{
\item \code{coc_number}: CoC number
\item \code{year}: Year
\item \code{eviction_filings}: Number of eviction court cases filed
\item \code{evictions}: Number of evictions
\item \code{eviction_filing_rate}: Number of eviction court cases filed / renting households
\item \code{eviction_rate}: Number of evictions / renting households
\item \code{missing_evictions_rate}: Share of the CoC's renting households in counties with missing eviction data
}
}
\description{
Builds counts and rates of eviction and eviction court case filing at the CoC level.
The rates are the number of evictions/court cases divided by the number of renting
households in the CoC.
}
