% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_fmr.R
\name{process_split_county_fmr}
\alias{process_split_county_fmr}
\alias{build_coc_fmr}
\title{Builds CoC level FMRs}
\usage{
process_split_county_fmr(processed_fmr, acs)

build_coc_fmr(processed_fmr, acs, crosswalk)
}
\arguments{
\item{processed_fmr}{Data frame with processed FMRs}

\item{acs}{Data frame with the county subdivision renter household counts}

\item{crosswalk}{Data frame with the county to CoC crosswalk}
}
\value{


A data frame with CoC level FMRs.
}
\description{
Calculates a CoC level FMR by taking a weighted average of the county FMRs in
the CoC. The weights are the share of the CoC population from each county.
}
\section{Functions}{
\itemize{
\item \code{process_split_county_fmr}: Convert sub-county FMRs to county FMRs

Calculates a weighted average FMR for the counties that are in more than one
FMR region. County subdivisions are unique to an FMR area so we take the
weighted average of FMRs in the county subdivisions to get the county level
FMR.
}}

