% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{build_coc_income}
\alias{build_coc_income}
\title{Creates CoC average median income measures}
\usage{
build_coc_income(yr, county_crosswalk, income_variables)
}
\arguments{
\item{yr}{A numeric with the year of the data.}

\item{county_crosswalk}{A data frame with the county to CoC crosswalk.}

\item{income_variables}{A character vector (optionally named with variable names)
with the ACS variable codes.}
}
\value{
A data frame:
\itemize{
\item \code{coc_number}: CoC_Number
\item \code{year}: Year
\item Variables: Weighted average of the provided variables
(weights are county share of the CoC population)
}
}
\description{
Takes the weighted average of the provided income measures, where the weights are
the county's share of the total CoC population. The default measures are: median
household income, median family income, and median individual earnings (pay from
wages and salary).
}
