% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{build_coc_list}
\alias{build_coc_list}
\title{Build list of CoCs}
\usage{
build_coc_list(coc_shapefile)
}
\arguments{
\item{coc_shapefile}{An sf data frame with the shapefile data.}
}
\value{
A data frame:
\itemize{
\item coc_number: CoC number
\item coc_name: CoC name
\item year: Year
}
}
\description{
Creates the list of CoCs and their names by year.
}
