% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_renter_share.R
\name{build_coc_rent_burden}
\alias{build_coc_rent_burden}
\title{CoC-level shares of rent burdened households}
\usage{
build_coc_rent_burden(year, tract_crosswalk)
}
\arguments{
\item{tract_crosswalk}{A data frame with a census tract to CoC crosswalk from
\code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}.}

\item{yr}{A numeric with the year of the data.}
}
\value{
A data frame:
\itemize{
\item \code{coc_number}: CoC number (character)
\item \code{year}: Year (numeric)
\item \code{share_rent_over_30_pct_inc}: Share of renter households paying more than
30 percent of their income in rent (numeric)
\item \code{share_rent_over_50_pct_inc}: Share of renter households paying more than
50 percent of their income in rent (numeric)
\item \code{median_rent_burden}: Median rent share of income in the CoC (numeric)
}
}
\description{
Builds shares of renter households who pay more than 30 or 50 percent of
their income in rent and the average median rent burden
(rent / household income) in the CoC.
}
