% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{build_coc_rent_burdened_share}
\alias{build_coc_rent_burdened_share}
\title{CoC-level shares of rent burdened households}
\usage{
build_coc_rent_burdened_share(county_rent_data, county_crosswalk)
}
\arguments{
\item{county_rent_data}{A data frame of county level counts of rent burdened
households created by \code{\link[=get_county_rent_burdened_count]{get_county_rent_burdened_count()}}.}

\item{county_crosswalk}{A data frame with a county to CoC crosswalk created
by \code{\link[=build_county_crosswalk]{build_county_crosswalk()}}.}
}
\value{
A data frame:
\itemize{
\item \code{year}: Year (numeric)
\item \code{coc_number}: CoC number (character)
\item \code{share_rent_over_30_pct_inc}: Share of renter households paying more than
30 percent of their income in rent (numeric)
\item \code{share_rent_over_50_pct_inc}: Share of renter households paying more than
50 percent of their income in rent (numeric)
}
}
\description{
Builds shares of renter households who pay more than 30 or 50 percent of
their income in rent.
}
\seealso{
\code{\link[=get_county_rent_burdened_count]{get_county_rent_burdened_count()}} for fetching the ACS county-level
rent burdened counts
}
