% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{build_county_renter_share}
\alias{build_county_renter_share}
\alias{build_coc_renter_shares}
\title{Calculates the share of renters in each CoC}
\usage{
build_county_renter_share(year)

build_coc_renter_shares(renter_shares, crosswalk)
}
\arguments{
\item{year}{}

\item{renter_shares}{A data frame with county renter shares}

\item{crosswalk}{A data frame with a county to CoC crosswalk}
}
\value{


A data frame with CoC renter shares.
}
\description{
Calculates the share of renters in each CoC
}
\section{Functions}{
\itemize{
\item \code{build_county_renter_share}: Calculates the share of renters in each county
}}

