% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{build_coc_unemployment}
\alias{build_coc_unemployment}
\title{CoC-level unemployment rates}
\usage{
build_coc_unemployment(unemployment_data, county_crosswalk)
}
\arguments{
\item{unemployment_data}{A data frame with raw unemployment data from
\code{\link[=fetch_unemployment]{fetch_unemployment()}}.}

\item{county_crosswalk}{A data frame with a county to CoC crosswalk from
\code{\link[=build_county_crosswalk]{build_county_crosswalk()}}.}
}
\value{
A data frame:
\itemize{
\item \code{coc_number}: CoC number
\item \code{year}: Year
\item \code{unemployment_rate}: CoC unemployment rate
}
}
\description{
Builds CoC level unemployment rates by taking a weighted average of county
unemployment rates from the \href{https://www.bls.gov/lau/}{BLS Local Area Unemployment Statistics (LAUS)}. The weights are the share of the CoC's
population coming from each county.
}
\seealso{
\code{\link[=fetch_unemployment]{fetch_unemployment()}} for downloading the BLS data
}
