% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{build_coc_vacancy_rate}
\alias{build_coc_vacancy_rate}
\title{CoC rental housing vacancy rents}
\usage{
build_coc_vacancy_rate(yr, county_crosswalk)
}
\arguments{
\item{yr}{A numeric with the year of the data.}

\item{county_crosswalk}{A data frame with a county to CoC crosswalk from
\code{\link[=build_county_crosswalk]{build_county_crosswalk()}}.}
}
\value{
A data frame with the rental vacancy rates:
\itemize{
\item \code{coc_number}: CoC number (character)
\item \code{year}: Year (numeric)
\item \code{rental_vacancy_rate}: Share of rental housing units not rented (numeric)
}
}
\description{
Builds rental housing vacancy rates from ACS table DP04.
}
