% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{build_coc_vacancy_rates}
\alias{build_coc_vacancy_rates}
\title{CoC housing vacancy rents}
\usage{
build_coc_vacancy_rates(yr, tract_crosswalk)
}
\arguments{
\item{yr}{A numeric with the year of the data.}

\item{tract_crosswalk}{A data frame with a census tract to CoC crosswalk from
\code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}.}
}
\value{
A data frame with the rental vacancy rates:
\itemize{
\item \code{coc_number}: CoC number (character)
\item \code{year}: Year (numeric)
\item \code{gross_vacancy_rate}: Share of all housing units that are unoccupied
(numeric)
\item \code{rental_vacancy_rate}: Share of rental housing units not rented (numeric)
}
}
\description{
Builds gross and rental vacancy rates.
}
\section{Definitions}{
Gross vacancy rate = all unoccupied housing units / all housing units

Rental vacancy rate = unoccupied rental units that are for rent /
(occupied rental units + unoccupied rental units that are for rent +
rented units that are unoccupied)
}

