% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_zillow_rent.R
\name{build_coc_zillow_rent}
\alias{build_coc_zillow_rent}
\title{Creates a CoC level Zillow rent index}
\usage{
build_coc_zillow_rent(tract_rent, tract_to_coc)
}
\arguments{
\item{tract_rent}{A data frame with processed census tract level Zillow rent
index data created by \code{\link[=build_tract_zillow_rent]{build_tract_zillow_rent()}}.}

\item{tract_to_coc}{A data frame with a census tract to CoC crosswalk created
by \code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}.}
}
\value{
A data frame with the CoC level Zillow rent index.
\itemize{
\item \code{coc_number}: CoC number (character)
\item \code{year}: year (integer)
\item \code{coc_rent_zillow}: annual average Zillow rent index, NA for CoC
with no zip codes covered by the Zillow rent index (numeric)
\item \code{coc_share_na_rent}: share of the CoC's population in zip codes not
covered by the Zillow rent index (numeric)
}
}
\description{
The CoC level index is created by taking a weighted average of the census
tract level Zillow rent index for the tracts in each CoC. The weights are the
share of the CoC's population that are in each tract.
}
\seealso{
\code{\link[=process_zillow_data]{process_zillow_data()}} for processing the raw Zillow rent index,
\code{\link[=build_coc_zillow_rent]{build_coc_zillow_rent()}} for creating the CoC level Zillow rent index,
\code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}} for creating the census tract to CoC crosswalk
}
