% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_display_map}
\alias{build_display_map}
\title{Moves Alaska and Hawaii for compact 50 state map}
\usage{
build_display_map(
  shapefile,
  smooth = 0.005,
  rotate_ak = -27,
  scale_ak = 0.4,
  shift_ak = c(-5e+05, -3250000),
  rotate_hi = -25,
  scale_hi = 1,
  shift_hi = c(5e+06, -1400000)
)
}
\arguments{
\item{shapefile}{A shapefile with CoC boundaries.}

\item{smooth}{A numeric parameter for shapefile simplification.}

\item{rotate_ak}{Rotation parameter for Alaksa, in degrees.}

\item{scale_ak}{Scale factor for Alaska.}

\item{shift_ak}{A vector with the horizonal and vertical shift for Alaska.}

\item{rotate_hi}{Rotation parameter for Hawaii, in degrees}

\item{scale_hi}{Scale factor for Hawaii.}

\item{shift_hi}{A vector with the horizonal and vertical shift for Hawaii}
}
\value{
A spatial data frame with adjusted representations of Alaska and Hawaii.
}
\description{
Moves, scales, and rotates Alaska and Hawaii to make a compact 50 state map for
data vizualization.
}
