% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{build_regex}
\alias{build_regex}
\title{Build file extension regular expression}
\usage{
build_regex(file_extensions = c("gdb", "shp"))
}
\arguments{
\item{file_extensions}{A character vector of shapefile file extensions to
include in the regular expression.}
}
\value{
A character vector containing the regular expression.
}
\description{
Creates a regular expression used to filter shapefile in input_data
}
\keyword{internal}
