% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_zillow_rent.R
\name{build_tract_zillow_rent}
\alias{build_tract_zillow_rent}
\title{Creates a census tract level Zillow rent index}
\usage{
build_tract_zillow_rent(rent_data, tract_to_zip)
}
\arguments{
\item{rent_data}{A data frame with processed zip code level Zillow rent index
data created by \code{\link[=process_zillow_data]{process_zillow_data()}}}

\item{tract_to_zip}{A data frame with a processed USPS census tract to zip
code crosswalk created by \code{\link[=process_tract_to_zip]{process_tract_to_zip()}}.}
}
\value{
A data frame with the census tract level Zillow rent index.
\itemize{
\item \code{tract_fips}: census tract FIPS code (character)
\item \code{year}: year (integer)
\item \code{tract_annual_mean_rent}: annual average Zillow rent index, NA for tracts
with no zip codes covered by the Zillow rent index (numeric)
\item \code{tract_share_na_rent}: share of the tract's residence in zip codes not
covered by the Zillow rent index (numeric)
}
}
\description{
The census tract level index is created by taking a weighted average of the
original zip code level Zillow rent index for the zip codes in each census
tract. The weights are the share of the census tract's residences that are in
each zip code. That data comes from the USPS census tract to zip code
crosswalk.
}
\seealso{
\code{\link[=process_zillow_data]{process_zillow_data()}} for processing the raw Zillow rent index,
\code{\link[=process_tract_to_zip]{process_tract_to_zip()}} for processing the census tract to zip code
crosswalk, \code{\link[=build_coc_zillow_rent]{build_coc_zillow_rent()}} for creating the CoC level Zillow
rent index
}
