% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_county_tract_crosswalk.R
\name{clip_tracts}
\alias{clip_tracts}
\title{Trim tracts outside of CoC boundaries}
\usage{
clip_tracts(tract_geodata, dissolved_cocs)
}
\arguments{
\item{tract_geodata}{An sf object with the tract data & boundaries}

\item{dissolved_cocs}{An sf object with the internal boundaries dissolved
since we only care at this point if a tract is in \emph{any} CoC}
}
\value{
A list with two sf objects: one with the tracts that are in a CoC and
one with the tracts that aren't in a CoC
}
\description{
Not all Census tracts are in CoC boundaries, this separates out the tracts
that are not in a CoC.
}
