% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{download_shapefile}
\alias{download_shapefile}
\title{Download a state's CoC shapefiles}
\usage{
download_shapefile(state_abb, state_name, year, td = tempdir())
}
\arguments{
\item{state_abb}{A character vector with two letter state abbreviations.}

\item{state_name}{A character vector with state names.}

\item{year}{A numeric with the year of the shapefiles.}

\item{td}{A directory where the shapefiles will be unzipped. Defaults to the
session temporary directory.}
}
\value{
A spatial data frame:
\itemize{
\item coc_number: CoC number
\item coc_name: CoC name
\item year: Year
\item geometry: The CoC's spatial data
}
}
\description{
Downloads the CoC shapefiles for a state in the given year.
}
\seealso{
\code{\link[=get_pre_2013_coc_shapefiles]{get_pre_2013_coc_shapefiles()}} for the main function
}
\keyword{internal}
