% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fetch_acs}
\alias{fetch_acs}
\title{Fetches ACS table data from the Census Bureau API}
\usage{
fetch_acs(..., quiet = TRUE)
}
\arguments{
\item{...}{Parameters passed to \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.}

\item{quiet}{Boolean to suppress status messages.}
}
\value{
A data frame with the requested ACS data.
\itemize{
\item \code{year}: Year (numeric)
\item \code{fips}: Geographic unit FIPS code (character)
\item Additional columns with requested data
}
}
\description{
A wrapper around \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} which fetches data from the Census
Bureau API, adds a column with the year of the data, and removes the unneeded
margin of error columns.
}
\seealso{
\code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}
}
