% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_renter_share.R
\name{fetch_acs_rent_burden}
\alias{fetch_acs_rent_burden}
\title{Census Tract level counts of rent-burdened households}
\usage{
fetch_acs_rent_burden(state, year)
}
\arguments{
\item{year}{A numeric with the year of ACS data to fetch.}
}
\value{
A data frame:
\itemize{
\item \code{year}: Year (numeric)
\item \code{tract_fips}: Census Tract FIPS code (character)
\item \code{count_30_plus}: Count of renting households paying more than 30\% of their
income in rent (numeric)
\item \code{count_50_plus}: Count of renting households paying more than 50\% of their
income in rent (numeric)
\item \code{total_computed}: Count of renting households that have a calculated rent
share of income (numeric)
\item \code{median_rent_burden}: Median rent share of income in the county (numeric)
}
}
\description{
Gets ACS data that has the counts of households who pay more than 30 or 50
percent of their income in rent and the median rent burden
(rent / household income) in the tract.
}
\seealso{
\code{\link[=build_coc_rent_burden]{build_coc_rent_burden()}} for the main function
}
\keyword{internal}
