% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{fetch_acs_rental_vacancy_rate}
\alias{fetch_acs_rental_vacancy_rate}
\title{Fetches ACS rental vacancy rates}
\usage{
fetch_acs_rental_vacancy_rate(yr)
}
\arguments{
\item{yr}{A numeric with the year of the data.}
}
\value{
A data frame with county vacancy rates:
\itemize{
\item \code{fips}: County FIPS code (character)
\item \code{year}: Year (numeric)
\item \verb{rental_vacancy rate}: Share of rental housing units not rented (numeric)
}
}
\description{
Fetches ACS rental vacancy rates
}
\seealso{
\code{\link[=build_coc_vacancy_rate]{build_coc_vacancy_rate()}}
}
\keyword{internal}
