% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{fetch_acs_vacancy_data}
\alias{fetch_acs_vacancy_data}
\title{Fetches ACS vacancy data}
\usage{
fetch_acs_vacancy_data(state, year)
}
\value{
A data frame with tract vacancy counts:
\itemize{
\item \code{fips}: County FIPS code (character)
\item \code{year}: Year (numeric)
\item \code{total_housing_units}: Total housing units (numeric)
\item \code{vacant_housing_units}: Total vacant housing units (numeric)
\item \code{occupied_rental_units}: Occupied rental housing units (numeric)
\item \code{for_rent}: Rental housing units that are unoccupied and for rent (numeric)
\item \code{rented_not_occupied}: Rental housing units that have been rented but are
unoccupied (numeric)
}
}
\description{
Fetches ACS vacancy data
}
\seealso{
\code{\link[=build_coc_vacancy_rates]{build_coc_vacancy_rates()}}
}
\keyword{internal}
