% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{fetch_public_program_use}
\alias{fetch_public_program_use}
\title{Fetch ACS public program use data}
\usage{
fetch_public_program_use(year)
}
\arguments{
\item{year}{A numeric with the year of the data to fetch.}
}
\value{
A data frame:
\itemize{
\item \code{fips}: Census Tract FIPS code
\item \code{year}: Year
\item \code{total_hh_snap}: Total number of households from the SNAP enrollment table
\item \code{hh_with_snap}: Number of households receiving SNAP benefits
\item \code{total_hh_pub_assist}: Total number of households from the public
assistance enrollment table
\item \code{hh_with_pub_assist}: Number of households receiving public assistance
benefits
\item \code{total_hh_ssi}: Total number of households from the SSI enrollment table
\item \code{hh_with_ssi}: Number of households receiving SSI benefits
\item \code{total_male_19_64}: Total number of males age 19-64
\item \code{male_19_64_with_medicaid}: Number of males age 19-64 enrolled in Medicaid
\item \code{total_female_19_64}: Total number of females age 19-64
\item \code{female_19_64_with_medicaid}: Number of females age 19-64 enrolled in
Medicaid
}
}
\description{
Fetches Census Tract level enrollment in SNAP, public assistance, SSI, and
Medicaid from the Census API. Medicaid data is for age 19-64.
}
\seealso{
\code{\link[=build_coc_public_program_use]{build_coc_public_program_use()}} for CoC public program utilization
rates
}
