% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{fetch_unemployment}
\alias{fetch_unemployment}
\title{Fetch raw BLS unemployment rates}
\usage{
fetch_unemployment(url, min_year = 2010)
}
\arguments{
\item{url}{A string with the url of the flat text file.}

\item{min_year}{A number with the first year of data to keep.}
}
\value{
A data frame:
\itemize{
\item \code{series}: BLS series ID
(see https://download.bls.gov/pub/time.series/la/la.txt)
\item \code{year}: Year
\item \code{month}: Month of the data (annual averages are "M13")
\item \code{value}: Unemployment rate
\item \code{footnote}: BLS footnote code
(see https://download.bls.gov/pub/time.series/la/la.footnote)
}
}
\description{
Downloads flat text file with BLS unemployment data. Parses just unemployment
rate data from the download.
}
\seealso{
\code{\link[=build_coc_unemployment]{build_coc_unemployment()}} for building CoC-level unemployment rates
}
