% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_fmr.R
\name{get_states_with_split_counties}
\alias{get_states_with_split_counties}
\alias{get_acs_county_sub}
\title{Fetches ACS renter household counts}
\usage{
get_states_with_split_counties(processed_fmr)

get_acs_county_sub(year, processed_fmr)
}
\arguments{
\item{processed_fmr}{Data frame with processed FMRs}

\item{year}{Year of the ACS data}
}
\value{


A data frame with renter household counts.
}
\description{
Finds the states that have sub-county FMR areas and fetches renter household
counts in those states at the county subdivision level.
}
\section{Functions}{
\itemize{
\item \code{get_states_with_split_counties}: Finds the states with sub-county FMRs
}}

