% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_data_processing.R
\name{get_coc_categories}
\alias{get_coc_categories}
\title{Creates a table of urban/rural CoC categories}
\usage{
get_coc_categories(wide_pit_data)
}
\arguments{
\item{wide_pit_data}{A data frame with the PIT counts in wide form.}
}
\value{
A data frame with the urban/rural status for each CoC.
}
\description{
Creates a table of urban/rural CoC categories
}
