% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_data_processing.R
\name{get_long_pit_data}
\alias{get_long_pit_data}
\title{Reshapes the PIT count data to long form}
\usage{
get_long_pit_data(wide_pit_data)
}
\arguments{
\item{wide_pit_data}{A data frame with the PIT counts in wide form}
}
\value{
A data frame with the processed PIT counts in long form.
}
\description{
Puts the PIT counts in long form with the count categories in the \code{category}
column and the count values in the \code{count} column.
}
