% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_rates.R
\name{get_overall_pit_counts}
\alias{get_overall_pit_counts}
\title{Filter overall homelessness counts from PiT data}
\usage{
get_overall_pit_counts(pit_data)
}
\arguments{
\item{pit_data}{A data frame with PiT counts in long form from
\code{\link[=get_long_pit_data]{get_long_pit_data()}}.}
}
\value{
A data frame with just the overall homelessness counts:
\itemize{
\item \code{coc_number}: CoC number (character)
\item \code{year}: Year (numeric)
\item \code{overall_homeless}: Count of all unhoused people in the CoC (numeric)
}
}
\description{
Filter overall homelessness counts from PiT data
}
\keyword{internal}
