% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_data_processing.R
\name{get_pit_years}
\alias{get_pit_years}
\title{Parse the years covered by the PIT counts}
\usage{
get_pit_years(filepath)
}
\arguments{
\item{filepath}{Path to the PIT count excel file}
}
\value{
A character vector with the years in the file.
}
\description{
Analyzes the PIT count excel file to pull out the years for which the file
has data. It expects a file where each year is in its own sheet and the name
of the sheet is just the year.
}
