% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{get_pre_2013_shapefiles}
\alias{get_pre_2013_shapefiles}
\title{Fetches pre-2013 CoC shapefiles}
\usage{
get_pre_2013_shapefiles(year)
}
\arguments{
\item{year}{A numeric with the year of the shapefiles.}
}
\value{
A spatial data frame:
\itemize{
\item coc_number: CoC number
\item coc_name: CoC name
\item year: Year
\item geometry: The CoC's spatial data
}
}
\description{
Before 2013, HUD released individual shapefiles for each CoC instead of one
national file. This downloads the individual shapefiles and combines them.
}
\seealso{
\code{\link[=get_shapefiles]{get_shapefiles()}}
}
