% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{get_shapefiles}
\alias{get_shapefiles}
\title{Read in CoC shapefiles}
\usage{
get_shapefiles(year, raw_coc_shapefiles, crs)
}
\arguments{
\item{year}{A numeric with the shapefile year.}

\item{raw_coc_shapefiles}{A character vector of paths to shapefiles saved on
disk.}

\item{crs}{A numeric with the coordinate reference system (CRS) to use for
the shapefiles.}
}
\value{
A spatial data frame:
\itemize{
\item coc_number: CoC number
\item coc_name: CoC name
\item year: Year
\item geometry: The CoC's spatial data in the provided CRS
}
}
\description{
Reads in all CoC shapefiles. Pre-2013 shapefiles are downloaded from HUD,
2013+ shapefiles are read from \code{input_data}.
}
\details{
Prior to 2013, HUD released individual shapefiles for each CoC and grouped
them by state. This function handles downloading and extracting those
shapefiles in \code{\link[=pre_2013_shapefile]{pre_2013_shapefile()}}.
}
