% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_zillow_rent.R
\name{get_tract_to_zip_year}
\alias{get_tract_to_zip_year}
\title{Parses year from tract to zip crosswalk}
\usage{
get_tract_to_zip_year(file_path)
}
\arguments{
\item{file_path}{A character vector with the path to the crosswalk.}
}
\value{
An integer value with the year of the crosswalk.
}
\description{
Parses year from tract to zip crosswalk
}
\seealso{
\code{\link[=process_tract_to_zip]{process_tract_to_zip()}}
}
\keyword{internal}
