% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_renter_share.R
\name{make_coc_rent_burden}
\alias{make_coc_rent_burden}
\title{Construct CoC-level shares of rent burdened households}
\usage{
make_coc_rent_burden(tract_rent_data, tract_crosswalk, yr)
}
\arguments{
\item{tract_rent_data}{A data frame of tract level counts of rent burdened
households created by \code{\link[=get_county_rent_burdened_count]{get_county_rent_burdened_count()}}.}

\item{tract_crosswalk}{A data frame with a tract to CoC crosswalk created
by \code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}.}
}
\value{
A data frame:
\itemize{
\item \code{year}: Year (numeric)
\item \code{coc_number}: CoC number (character)
\item \code{share_rent_over_30_pct_inc}: Share of renter households paying more than
30 percent of their income in rent (numeric)
\item \code{share_rent_over_50_pct_inc}: Share of renter households paying more than
50 percent of their income in rent (numeric)
\item \code{median_rent_burden}: Median rent share of income in the CoC (numeric)
}
}
\description{
Builds shares of renter households who pay more than 30 or 50 percent of
their income in rent and the average median rent burden
(rent / household income) in the CoC.
}
\seealso{
\code{\link[=build_coc_rent_burden]{build_coc_rent_burden()}} for the main function
}
\keyword{internal}
