% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_renter_share.R
\name{make_coc_rental_vacancy_rate}
\alias{make_coc_rental_vacancy_rate}
\title{Constructs the CoC rental vacancy rates}
\usage{
make_coc_rental_vacancy_rate(acs_data, yr, county_crosswalk)
}
\arguments{
\item{acs_data}{A data frame with county vacancy rates from
\code{\link[=fetch_acs_rental_vacancy_rate]{fetch_acs_rental_vacancy_rate()}}}

\item{yr}{A numeric with the year of the data.}

\item{county_crosswalk}{A data frame with a county to CoC crosswalk from
\code{\link[=build_county_crosswalk]{build_county_crosswalk()}}.}
}
\description{
Constructs the CoC rental vacancy rates
}
\seealso{
\code{\link[=build_coc_vacancy_rate]{build_coc_vacancy_rate()}}
}
\keyword{internal}
