% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{make_coc_unemployment}
\alias{make_coc_unemployment}
\title{Construct the CoC unemployment rate}
\usage{
make_coc_unemployment(unemployment_data, county_crosswalk)
}
\arguments{
\item{unemployment_data}{A data frame with raw unemployment data from
\code{\link[=fetch_unemployment]{fetch_unemployment()}}.}

\item{county_crosswalk}{A data frame with a county to CoC crosswalk from
\code{\link[=build_county_crosswalk]{build_county_crosswalk()}}.}
}
\value{
A data frame:
\itemize{
\item \code{coc_number}: CoC number
\item \code{year}: Year
\item \code{unemployment_rate}: CoC unemployment rate
}
}
\description{
Construct the CoC unemployment rate
}
\seealso{
\code{\link[=build_coc_unemployment]{build_coc_unemployment()}} for the main function
}
\keyword{internal}
