% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_rates.R
\name{make_pit_rates}
\alias{make_pit_rates}
\title{Constructs the homelessness rates}
\usage{
make_pit_rates(pit_data, coc_populations)
}
\arguments{
\item{pit_data}{A data frame with PiT counts in long form from
\code{\link[=get_long_pit_data]{get_long_pit_data()}}.}

\item{coc_populations}{A data frame with CoC overall populations and
populations in poverty from \code{\link[=build_coc_populations]{build_coc_populations()}}.}
}
\description{
Constructs the homelessness rates
}
\keyword{internal}
