% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{match_tract_to_coc}
\alias{match_tract_to_coc}
\alias{recombine_tracts}
\alias{map_tracts_to_cocs}
\title{Match tracts to CoCs}
\usage{
match_tract_to_coc(tract_points, coc_boundaries)

recombine_tracts(tracts_in_cocs, tracts_not_in_cocs)

map_tracts_to_cocs(clipped_tracts, coc_boundaries)
}
\arguments{
\item{tract_points}{sf object with the center points of the Census Tracts}

\item{coc_boundaries}{An sf object with CoC boundaries}

\item{tracts_in_cocs}{A data frame with tracts matched to CoCs}

\item{tracts_not_in_cocs}{A data frame with tracts not in CoCs}

\item{clipped_tracts}{A list with two elements: \code{tracts_in_cocs} &
\code{tracts_not_in_cocs}}
}
\value{
A clean data frame of all tracts, their tract population data, and
the corresponding CoC if there is one.
}
\description{
Calls functions that calculate the center points for tracts in CoCs, match
them to the CoC the center point lies within, recombine the tracts in and not
in CoCs, and does some light cleaning.
}
\section{Functions}{
\itemize{
\item \code{match_tract_to_coc}: Match Census Tracts to CoCs

Matches Census Tracts to CoCs based on the CoC the center of the Census Tract
lies within.

\item \code{recombine_tracts}: Combine tables of tracts in and not in CoCs

Takes the matched table of tracts in CoCs, joins it with the table of tracts
not in CoCs and cleans up the variables.
}}

