% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{match_tract_to_coc}
\alias{match_tract_to_coc}
\title{Match Census Tracts to CoCs}
\usage{
match_tract_to_coc(tract_points, coc_boundaries)
}
\arguments{
\item{tract_points}{sf object with the center points of the Census Tracts}

\item{coc_boundaries}{sf object with the CoC boundaries}
}
\value{
A data frame with the tract, tract data, and its corresponding CoC
}
\description{
Matches Census Tracts to CoCs based on the CoC the center of the Census Tract
lies within.
}
