% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{move_state}
\alias{move_state}
\title{Move, scale, rotate a state on a map}
\usage{
move_state(map, state, rotation = 0, scale = 1, shift = c(0, 0))
}
\arguments{
\item{map}{A spatial data frame with the map data.}

\item{state}{A string with the abbreviation of the state to adjust (ie. "AK").}

\item{rotation}{A numeric with rotation adjustment given in degrees.}

\item{scale}{A numeric with the scale factor for the state.}

\item{shift}{A numeric vector with the horizontal and vertical adjustments for
the state's position.}
}
\value{
The same spatial data frame but with the adjusted state.
}
\description{
Helps move states like Alaska and Hawaii so the map-based visualizations can be
more compact. This function simplifies changing the position of the state
(shifting), changing the size of the state (scaling), and rotating the state.
}
\keyword{internal}
