% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_data_processing.R
\name{parse_pit_year}
\alias{parse_pit_year}
\title{Read and clean raw PIT count data}
\usage{
parse_pit_year(filepath, year)
}
\arguments{
\item{filepath}{Path to the PIT count excel file}

\item{year}{Year of the data to parse}
}
\value{
A data frame with the processed PIT counts in wide form.
}
\description{
Reads in a tab on the pit spreadsheet and does some light cleaning by adding
a year column and making friendlier column names.
}
