% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_fmr.R
\name{process_fmr}
\alias{process_fmr}
\title{Read and process raw FMR data}
\usage{
process_fmr(raw_fmr_path)
}
\arguments{
\item{raw_fmr_path}{A character string with the file path of the raw FMR
spreadsheet}
}
\value{
A data frame with the processed FMR data.
}
\description{
Reads in the raw FMR spreadsheets, standardizes the FIPS code and FMR
variables, removes unneeded variables, and counts the number of records per
county.
}
\details{
Counting the records per county is necessary because unlike the rest
of the U.S, which uses groups of counties for FMR areas, New England states
create FMR areas with county subdivisions, so a county can be in more than
one FMR. This flags those records so we can deal with them further down the
line.
}
