% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_zillow_rent.R
\name{process_tract_to_zip}
\alias{process_tract_to_zip}
\title{Reads and processes tract to zip code crosswalk}
\usage{
process_tract_to_zip(file_path)
}
\arguments{
\item{file_path}{A character vector with the path to the crosswalk.}
}
\value{
A data frame with the processed crosswalk.
\itemize{
\item \code{zip}: zip code (character)
\item \code{tract_fips}: census tract FIPS code (character)
\item \code{year}: year (integer)
\item \code{res_ratio}: share of census tract's residential addresses in given zip
code (numeric)
}
}
\description{
Reads in the USPS census tract to zip code crosswalk, cleans up the variable
names and types, adds a year variable, and removes entries with no
residential units.
}
