% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_economic_indicators.R
\name{process_unemployment}
\alias{process_unemployment}
\title{Clean unemployment rates}
\usage{
process_unemployment(unemployment_data, use_month = "M01")
}
\arguments{
\item{unemployment_data}{A data frame with raw unemployment data from
\code{\link[=fetch_unemployment]{fetch_unemployment()}}.}

\item{use_month}{A character value with the month to use or the annual
average. For annual average use "M13". Defaults to January ("M01").}
}
\value{
A data frame:
\itemize{
\item \code{county_fips}: County FIPS code
\item \code{year}: Year
\item \code{unemployment_rate}: Unemployment rate
}
}
\description{
Grabs the chosen month's or annual average unemployment rate and parses the
county FIPS code.
}
\seealso{
\code{\link[=build_coc_unemployment]{build_coc_unemployment()}} for the main function
}
\keyword{internal}
