% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_zillow_rent.R
\name{process_zillow_data}
\alias{process_zillow_data}
\title{Reads and processes raw Zillow rent index data}
\usage{
process_zillow_data(file_path)
}
\arguments{
\item{file_path}{A character string with the path to the raw data.}
}
\value{
A data frame with the annual average rent by zip code.
\itemize{
\item \code{zip}: zip code (character)
\item \code{year}: year (integer)
\item \code{annual_mean_rent}: average annual Zillow rent index (numeric)
}
}
\description{
Reads in the raw Zillow data, cleans up variable names and types, reshapes it
to long form, and calculates an annual average from the monthly data.
}
\seealso{
\code{\link[=build_coc_zillow_rent]{build_coc_zillow_rent()}} for creating the CoC level Zillow rent
index, \code{\link[=build_tract_zillow_rent]{build_tract_zillow_rent()}} for creating the census tract level
Zillow rent index
}
