% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/process_coc_shapefiles.R
\name{read_raw_coc_shapefile}
\alias{read_raw_coc_shapefile}
\title{Read in original CoC shapefile}
\usage{
read_raw_coc_shapefile(shapefile_path)
}
\arguments{
\item{shapefile_path}{Path to the shapefile or other spatial data
file/database}
}
\value{
The shapefile in a spatial data frame
}
\description{
Loads an original CoC shapefile from HUD as a spatial data frame and adds a
year column. Returns an error if it cannot parse a 4-digit year from the file
path.
}
\seealso{
\code{\link[=simplify_shapefile]{simplify_shapefile()}} for simplifying the shapefile to reduce its
complexity
}
