% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{recombine_tracts}
\alias{recombine_tracts}
\title{Combine tables of tracts in and not in CoCs}
\usage{
recombine_tracts(tracts_in_cocs, tracts_not_in_cocs)
}
\arguments{
\item{tracts_in_cocs}{A data frame with tracts matched to CoCs}

\item{tracts_not_in_cocs}{A data frame with tracts not in CoCs}
}
\value{
A clean data frame of all tracts, their tract population data, and
the corresponding CoC if there is one.
}
\description{
Takes the matched table of tracts in CoCs, joins it with the table of tracts
not in CoCs and cleans up the variables.
}
