% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_county_tract_crosswalk.R
\name{write_crosswalk}
\alias{write_crosswalk}
\title{Write the crosswalk to CSV}
\usage{
write_crosswalk(crosswalk, type, output_directory)
}
\arguments{
\item{crosswalk}{A crosswalk data frame}

\item{type}{A string with the type of crosswalk. Either \code{"tract"} or \code{"county"}.}

\item{output_directory}{A path to the directory to save the CSV}
}
\value{
Returns the CSV path so \link{targets} can track it
}
\description{
Creates a CSV flat text file with the crosswalk data.
}
