% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/helpers.R
\name{write_dataset}
\alias{write_dataset}
\title{Write combined dataset to disk}
\usage{
write_dataset(
  data,
  output_function,
  extension,
  output_dir = "output_data",
  file_name = "coc_data"
)
}
\arguments{
\item{data}{A data frame with the combined dataset.}

\item{output_function}{A character string with the function that will
actually write the file.}

\item{extension}{A character string with the extension for the output file.}

\item{output_dir}{A character string with the path to the output directory.}

\item{file_name}{A character string with the file name.}
}
\value{
Invisibly returns a character string with the output file's path.
}
\description{
Flexibly writes the combined data set to disk. Since the actual call to the
function that will write the dataset is built dynamically, it can support
many filetypes as long as they are of the form \verb{output_function(data, file_path}.
}
