% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{build_county_crosswalk}
\alias{build_county_crosswalk}
\title{Creates the county to CoC crosswalk}
\usage{
build_county_crosswalk(tract_crosswalk)
}
\arguments{
\item{tract_crosswalk}{A tract to CoC crosswalk}
}
\value{
A data frame with the CoC population,share of the CoC coming from
each county, and share of each CoC in each county.
}
\description{
Calculates the CoC population, share of the CoC coming from each tract, and share of each CoC in each county.
Both for the total population and the population under the poverty line.
}
\seealso{
\code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}
}
