% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/pit_rates.R
\name{build_pit_rates}
\alias{build_pit_rates}
\title{Calculate homelessness rates by CoC}
\usage{
build_pit_rates(pit_data, coc_populations)
}
\arguments{
\item{pit_data}{A data frame with PIT counts in long form}

\item{coc_populations}{A data frame with CoC overall populations and
populations in poverty}
}
\value{
A data frame with CoC homelessness counts, rates, and number of
unhoused people per 1000 people in the overall population.
}
\description{
Calculate homelessness rates by CoC
}
