% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_county_tract_crosswalk.R
\name{build_tract_crosswalk}
\alias{build_tract_crosswalk}
\title{Creates the tract to CoC crosswalk}
\usage{
build_tract_crosswalk(recombined_tracts)
}
\arguments{
\item{recombined_tracts}{A data frame of tracts, their populations, and the
corresponding CoC}
}
\value{
A data frame with the CoC population and share of the CoC coming from
each tract.
}
\description{
Calculates the CoC population and share of the CoC coming from each tract.
Both for the total population and the population under the poverty line.
}
