% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{fetch_tract_data}
\alias{fetch_tract_data}
\title{Gets ACS data from the Census API}
\usage{
fetch_tract_data(year, crs)
}
\arguments{
\item{year}{Year of the ACS data and tract boundaries to fetch}

\item{crs}{The EPSG code for the tract spatial data CRS}
}
\value{
An sf object with tract population data and boundaries
}
\description{
Requests ACS population data for all tracts in the 50 states + DC as well as
spatial data for the tracts. It also converts the spatial data to the given
CRS.
}
