% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_renter_share.R
\name{make_coc_vacancy_rates}
\alias{make_coc_vacancy_rates}
\title{Constructs the CoC vacancy rates}
\usage{
make_coc_vacancy_rates(acs_data, yr, tract_crosswalk)
}
\arguments{
\item{acs_data}{A data frame with tract vacancy data from
\code{\link[=fetch_acs_vacancy_data]{fetch_acs_vacancy_data()}}}

\item{tract_crosswalk}{A data frame with a census tract to CoC crosswalk from
\code{\link[=build_tract_crosswalk]{build_tract_crosswalk()}}.}
}
\description{
Constructs the CoC vacancy rates
}
\seealso{
\code{\link[=build_coc_vacancy_rates]{build_coc_vacancy_rates()}}
}
\keyword{internal}
