% Generated by roxygen2: do not edit by hand
% Please edit documentation in code/coc_county_tract_crosswalk.R
\name{map_tracts_to_cocs}
\alias{map_tracts_to_cocs}
\title{Match tracts to CoCs}
\usage{
map_tracts_to_cocs(clipped_tracts, coc_boundaries)
}
\arguments{
\item{clipped_tracts}{A list with two elements: \code{tracts_in_cocs} &
\code{tracts_not_in_cocs}}

\item{coc_boundaries}{An sf object with CoC boundaries}
}
\value{
A clean data frame of all tracts, their tract population data, and
the corresponding CoC if there is one.
}
\description{
Calls functions that calculate the center points for tracts in CoCs, match
them to the CoC the center point lies within, recombine the tracts in and not
in CoCs, and does some light cleaning.
}
