% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coc_county_tract_crosswalk.R
\name{match_tracts_to_cocs}
\alias{match_tracts_to_cocs}
\title{Match Census Tracts to CoCs}
\usage{
match_tracts_to_cocs(tract_points, coc_boundaries)
}
\arguments{
\item{tract_points}{sf object with the center points of the Census Tracts}

\item{coc_boundaries}{sf object with the CoC boundaries}
}
\description{
Matches Census Tracts to CoCs based on the CoC the center of the Census Tract
lies within.
}
