% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coc_shapefiles.R
\name{read_raw_coc_shapefile}
\alias{read_raw_coc_shapefile}
\title{Read in original CoC shapefile}
\usage{
read_raw_coc_shapefile(shapefile_path)
}
\arguments{
\item{shapefile_path}{Path to the shapefile or other spatial data
file/database}
}
\value{
A spatial data frame:
\itemize{
\item coc_number: CoC number
\item coc_name: CoC name
\item year: Year
\item geometry: The CoC's spatial data
}
}
\description{
Loads an original CoC shapefile from HUD as a spatial data frame and adds a
year column. Returns an error if it cannot parse a 4-digit year from the file
path.
}
\seealso{
\code{\link[=get_shapefiles]{get_shapefiles()}}
}
